Handle cases where `stale_checks` callback returns `nil` (#3)
Previously, returning `nil` for the `etag` or `last_modified` value in
the `stale_check` callback would lead to a 304 NOT MODIFIED if a user
requested the resource while providing a `if-none-match` or
`if-modified-since` header in their request. I wasn't the only one to
notice this. [This
PR](https://github.com/michalmuskala/phoenix_etag/pull/4) has been open
since 2017, but it wasn't a complete fix and doesn't include tests.

I have added a significant number of test cases as well as refactored
the implementation to be more logical to my brain. It was easier for me
to think about the narrower cases where the data was _fresh_ (they
provided the relevant header and the `stale_check` provided a matching
value), rather than the cases where it should be considered _stale_
(everything else).

We're definitely not fully compliant with
[RFC9110](https://www.rfc-editor.org/rfc/rfc9110#name-precedence-of-preconditions),
but it's an improvement.